Update notification-box
Reverse the list for the notifications. Now when a new notification is added, will be added on the top of the list. This is a more natural way to see the notification, the most recent stay always on the top. Change the static "10" string that show the number of notifications with che actual count of the notifications.